requirements.txt
discord.py
newsapi-python